python/twitter_api/requirements.txt
certifi==2018.4.16
chardet==3.0.4
future==0.16.0
idna==2.7
oauthlib==2.1.0
python-twitter==3.4.2
requests==2.19.0
requests-oauthlib==1.0.0
urllib3==1.23
